README links to the strawlab/Miriam repository
Thanks for all the work on Miriam! We think it is a very clever design
with great potential.

We have been working pretty heavily on it for the past months and have a
number of improvements and validations. Since several of the issues we
opened in the Miroculus repository were not addressed within a span of
weeks, and as agreed via email with Miroculus employees, we have been
working in our own repository. Since some of these issues would prevent
basic operation or even fabrication, it seems appropriate to redirect
interested people to our repository where these issues have been
addressed.